requirements.txt
streamlit==1.21.0
pymongo==4.3.3
bcrypt==3.2.0
openai==0.27.0
python-dotenv==1.0.0
google-generativeai==0.8.3  # You may need to adjust this based on available versions
